requirements.txt
PyYAML==6.0
fastapi==0.75.0
uvicorn==0.17.0
tortoise-orm==0.19.0
asyncpg==0.25.0